Move width code out
