+++ src/gen/lookup.rs
@@ -310,18 +310,6 @@ pub(crate) fn width_in_str(c: char, mut next_info: WidthInfo) -> (i8, WidthInfo)
     }
 }
 
-#[inline]
-pub fn str_width<S: DoubleEndedIterator<Item = char>>(s: S) -> usize {
-    s.rfold(
-        (0, WidthInfo::DEFAULT),
-        |(sum, next_info), c| -> (usize, WidthInfo) {
-            let (add, info) = width_in_str(c, next_info);
-            (sum.wrapping_add_signed(isize::from(add)), info)
-        },
-    )
-    .0
-}
-
 /// Returns the [UAX #11](https://www.unicode.org/reports/tr11/) based width of `c` by
 /// consulting a multi-level lookup table.
 ///
@@ -627,17 +615,4 @@ pub(crate) fn width_in_str_cjk(c: char, mut next_info: WidthInfo) -> (i8, WidthI
         let ret = lookup_width_cjk(c);
         (ret.0 as i8, ret.1)
     }
-}
-
-#[cfg(feature = "cjk")]
-#[inline]
-pub fn str_width_cjk<S: DoubleEndedIterator<Item = char>>(s: S) -> usize {
-    s.rfold(
-        (0, WidthInfo::DEFAULT),
-        |(sum, next_info), c| -> (usize, WidthInfo) {
-            let (add, info) = width_in_str_cjk(c, next_info);
-            (sum.wrapping_add_signed(isize::from(add)), info)
-        },
-    )
-    .0
 }
